Changes to stabilize MBW selection Bug fix for GLASSFISH-19417
Since we disabled the Jackson/JettisonFeatures, we added a MapWriter to
handle the work they were doing.  What was not caught, however, is that
RestCollection implemented the Map interface (for historical reasons).
This resulted in an unpredictable MBW selection at runtime, all
dependent on the order in which the JVM returned the classes to the
Jersey runtime. If MapWriter came up first, RestCollections were encoded
improperly.  This change removes the Map interface from RestCollection,
as it is no longer needed.