Further fix for 19419
The new test in "prepare" for identifying a dotted name reference to a property assumed that the "property" string identifying that the target is a property would not be the first part of the string, but that's actually permitted.  The GlassFish admin devtests revealed this.

Tests: the offending admin devtest