Fix for 19525
Before allowing a user to enable secure admin the system checks to make sure that there are no admin users in the default file realm with empty passwords, which would be a security risk.  That checking code failed if the user configures a non-default admin realm, such as LDAP, throwing a NPE.

This change skips the check if the admin realm is not the file realm.

Tests: QL