First check-in to convert to injected password alias stores.
This set of changes introduces the new contract for PasswordAliasStore as well as the implementation for it and a service instance for the domain password aliases.  

Any uses of password aliases should now use the new interfaces and contracts, rather than the PasswordAdapter class directly.  PasswordAdapter remains, because the new JCEKS password alias store implementation delegates to it and because it is currently still used for some other purposes.  

Tests: QL, admin dev tests