Allow the app client deployer to add the permissions.xml file to the downloaded files.
Java EE 7 allows the developer to add META-INF/permissions.xml to the EAR or to a stand-alone client.  This file will be needed by the ACC as it prepares the runtime environment.

These changes add that file to the top-level generated JAR file for the deployment, thus making it accessible via getResource at runtime by the ACC.

Tests: QL