Fix FindBugs error.
FindBugs complained because the logger field is not set in the constructor and is dereferenced from boostrapSecureAdminRemotely and reportFailure.  Yet the value is assigned very early in the execute method - the only non-private method - and all dereferences are in private methods invoked after that assignment is done, so the field will have a value by the time it is dereferenced.

Anyway, initializing it to null placates FindBugs.