Open PasswordAdapter lazily
As part of abstracting the PasswordAliasStore the JCEKS implementation initialized the underlying PasswordAdapter immediately.  This brought some work into server start-up when it was not necessarily so earlier.

This change defers the initialization of the PasswordAdapter until the first actual use to avoid adding work to server start-up.

Tests: QL