Fix for GLASSFISH-19606
The restAuthURL property was added to allow the console to look up the URL with which it needs to authenticate
for REST access. This URL is well-known and not changeable, so the console should be able to calculate it using
data from the admin listener. This change removes the property from the domain template and all references to
it from the system. It also modifies the console to calculate the URL.

Changes review by Tim Quinn.
All admin dev tests pass. Manual testing of console passed.