Additional fix for GLASSFISH-20160 - asadmin get command referring to a default configured item fails
An earlier check-in restored a long-ago removed bit of the domain.xml template for the default security audit-module, but that confused the upgrade logic that had been changed when the audit-module had been removed earlier.

This change fixes the upgrade problem.

Approved by Tom
Reviewed by Tom
Passed QL, the upgrade GlassFish admin devtest