Fix for FPP-1648 - CommandSecurityChecker should throw helpful exception if @AccessRequired resource refers to a field that is null at runtime
The @AccessRequired annotations can refer to a field on the command class using ${fieldName} in the resource name expression.  By the time the field name is evaluated as part of security checking the field should be non-null.  This change causes the CommandSecurityChecker to throw an IllegalArgumentException indicating that a field has a null value.  (Previously it would just cause a NullPointerException without much diagnostic help.)