Enhanced fix for GLASSFISH-20613
The earlier fix for this changed the externally-visible string by which the monitored element was exposed (replacing the left bracket with '_ARRAY_') so as to prevent a URI-unfriendly resource name (including the left bracket with no matching right one) from being used for authorization.

These changes restore the externally-visible string, substituting _ARRAY_ for the left bracket only internally in the resource name.

Passed QL, the ejb dev test which failed in 20510