Further refinement to the fix for GLASSFISH-20671
The <ssl> element can be a child either of <protocol> or <jmx-connector>.  The config listener did not distinguish, always assuming the parent of the <ssl> element was a <protocol> which caused problems.  The earlier fix for GLASSFISH-20671 used an incorrect test to detect if the parent was in fact a <protocol>.  

This change repairs that problem.

Tests: QL, RQ, failing web devtests